change the main card builder to use the same form as the bot for code reuse
