+++ handler/cardhandlers.go
@@ -4,7 +4,6 @@ import (
 	"encoding/csv"
 	"encoding/json"
 	"fmt"
-	"html"
 	"io"
 	"log"
 	"net/http"
@@ -18,6 +17,7 @@ import (
 	"vc_file_grouper/structout"
 	"vc_file_grouper/util"
 	"vc_file_grouper/vc"
+	"vc_file_grouper/wiki"
 )
 
 // CardHandler shows cards in order
@@ -178,7 +178,7 @@ func CardDetailHandler(w http.ResponseWriter, r *http.Request) {
 		len(evolutions),
 		len(amalgamations),
 	)
-	var firstEvo, lastEvo *vc.Card
+	var firstEvo *vc.Card
 	var evokeys []string // cache of actual evos for this card
 	for _, k := range vc.EvoOrder {
 		if evo, ok := evolutions[k]; ok {
@@ -192,24 +192,6 @@ func CardDetailHandler(w http.ResponseWriter, r *http.Request) {
 				)
 				firstEvo = evo
 			}
-			if k == "H" {
-				log.Printf("Found Card %s: (%d) Last Evo: %s: (%d)",
-					card.Name,
-					card.ID,
-					evo.Name,
-					evo.ID,
-				)
-				lastEvo = evo
-			}
-		}
-	}
-
-	var turnOverTo, turnOverFrom *vc.Card
-	if firstEvo.ID > 0 {
-		if firstEvo.TransCardID > 0 {
-			turnOverTo = firstEvo.EvoAccident()
-		} else {
-			turnOverFrom = firstEvo.EvoAccidentOf()
 		}
 	}
 
@@ -254,198 +236,14 @@ func CardDetailHandler(w http.ResponseWriter, r *http.Request) {
 
 	fmt.Fprintf(w, "<div style=\"clear:both;float:left\">Edit on the <a href=\"https://valkyriecrusade.fandom.com/wiki/%s?action=edit\">fandom</a>\n<br /></div>", cardName)
 	io.WriteString(w, "<div><textarea readonly=\"readonly\" style=\"width:100%;height:450px\">")
+
+	cardPage := wiki.CardPage{}
 	if card.IsClosed != 0 {
-		io.WriteString(w, "{{Unreleased}}")
+		cardPage.PageHeader = "{{Unreleased}}"
 	}
-	if card.CardSymbolID > 0 {
-		fmt.Fprintf(w, "{{Card\n|element = %s\n|rarity = %s\n|symbol = %s\n", card.Element(), card.MainRarity(), card.Symbol())
-	} else {
-		fmt.Fprintf(w, "{{Card\n|element = %s\n|rarity = %s\n", card.Element(), card.MainRarity())
-	}
+	cardPage.CardInfo.UpdateAll(firstEvo, avail)
 
-	skillMap := make(map[string]string)
-
-	skillEvoMod := ""
-	if firstEvo.Rarity()[0] == 'G' {
-		skillEvoMod = "g"
-	} else if len(firstEvo.Rarity()) > 2 && firstEvo.Rarity()[0] == 'X' {
-		skillEvoMod = "x"
-	} else if evo, ok := evolutions["A"]; ok && firstEvo.ID == evo.ID {
-		// if the first evo is the amalgamation evo...
-		skillEvoMod = ""
-	}
-
-	skillMap[skillEvoMod] = printWikiSkill(firstEvo.Skill1(), lastEvo.Skill1(), skillEvoMod)
-	skillMap[skillEvoMod+"2"] = printWikiSkill(firstEvo.Skill2(), lastEvo.Skill2(), skillEvoMod+"2")
-	skillMap[skillEvoMod+"3"] = printWikiSkill(firstEvo.Skill3(), lastEvo.Skill3(), skillEvoMod+"3")
-	skillMap[skillEvoMod+"3"] = printWikiSkill(firstEvo.Skill3(), lastEvo.Skill3(), skillEvoMod+"3")
-	skillMap[skillEvoMod+"t"] = printWikiSkill(firstEvo.ThorSkill1(), lastEvo.ThorSkill1(), skillEvoMod+"t")
-
-	// add amal skills as long as the first evo wasn't the amal
-	if evo, ok := evolutions["A"]; ok && firstEvo.ID != evo.ID {
-		aSkillName := evo.Skill1Name()
-		if aSkillName != firstEvo.Skill1Name() {
-			skillMap["a"] = printWikiSkill(evo.Skill1(), nil, "a")
-		}
-		if _, ok := evolutions["GA"]; ok {
-			skillMap["ga"] = printWikiSkill(evo.Skill1(), nil, "ga")
-		}
-		if _, ok := evolutions["XA"]; ok {
-			skillMap["xa"] = printWikiSkill(evo.Skill1(), nil, "xa")
-		}
-	}
-	// add awoken skills as long as the first evo wasn't awoken
-	if evo, ok := evolutions["G"]; ok && firstEvo.ID != evo.ID {
-		skillMap["g"] = printWikiSkill(evo.Skill1(), nil, "g")
-		skillMap["g2"] = printWikiSkill(evo.Skill2(), nil, "g2")
-		skillMap["g3"] = printWikiSkill(evo.Skill3(), nil, "g3")
-		skillMap["gt"] = printWikiSkill(evo.ThorSkill1(), nil, "gt")
-	}
-	// add rebirth skills as long as the first evo wasn't rebirth
-	if evo, ok := evolutions["X"]; ok && firstEvo.ID != evo.ID {
-		skillMap["x"] = printWikiSkill(evo.Skill1(), nil, "x")
-		skillMap["x2"] = printWikiSkill(evo.Skill2(), nil, "x2")
-		skillMap["x3"] = printWikiSkill(evo.Skill3(), nil, "x3")
-		skillMap["xt"] = printWikiSkill(evo.ThorSkill1(), nil, "xt")
-	}
-	// order that we want to print the skills
-	skillEvos := []string{"", "2", "3", "a", "g", "g2", "g3", "ga", "x", "x2", "x3", "xa", "t", "gt", "xt"}
-
-	// actually print the skills now...
-	for _, skillEvo := range skillEvos {
-		if val, ok := skillMap[skillEvo]; ok && val != "" {
-			io.WriteString(w, val)
-		}
-	}
-
-	//traverse evolutions in order
-	lenEvoKeys := len(evokeys)
-	fmt.Fprint(w, "<!-- Not all evolution stats here will be applicable. Please delete any that are not needed. -->\n")
-	for _, k := range evokeys {
-		evo := evolutions[k]
-		if k == "H" {
-			if evo.EvolutionRank >= 0 {
-				k = strconv.Itoa(evo.EvolutionRank)
-			} else if lenEvoKeys == 1 {
-				k = "1"
-			}
-		}
-		maxAtk, maxDef, maxSol := maxStats(evo, len(evolutions))
-		maxLvl := 0
-		if evo.CardRarity() != nil {
-			maxLvl = evo.CardRarity().MaxCardLevel
-		}
-		fmt.Fprintf(w, "|max level %[1]s = %[2]d\n|cost %[1]s = %[3]d\n|atk %[1]s = %[4]d%s\n|def %[1]s = %[6]d%s\n|soldiers %[1]s = %[8]d%s\n",
-			strings.ToLower(k),
-			maxLvl,
-			evo.DeckCost,
-			evo.DefaultOffense, maxAtk,
-			evo.DefaultDefense, maxDef,
-			evo.DefaultFollower, maxSol)
-	}
-
-	for _, k := range evokeys {
-		evo := evolutions[k]
-		if k == "H" {
-			if evo.EvolutionRank >= 0 {
-				k = strconv.Itoa(evo.EvolutionRank)
-			} else if lenEvoKeys == 1 {
-				k = "1"
-			}
-		}
-		if evo.MedalRate > 0 {
-			fmt.Fprintf(w, "|medals %[1]s = %[2]d\n", strings.ToLower(k), evo.MedalRate)
-		}
-
-		if evo.Price > 0 {
-			fmt.Fprintf(w, "|gold %[1]s = %[2]d\n", strings.ToLower(k), evo.Price)
-		}
-	}
-
-	fmt.Fprintf(w, "|friendship points = %d\n", card.Character().MaxFriendship)
-
-	fmt.Fprintf(w, "|description = %s\n|friendship = %s\n",
-		html.EscapeString(card.Description()), html.EscapeString(strings.ReplaceAll(card.Friendship(), "\n", "<br />")))
-	login := card.Login()
-	if len(strings.TrimSpace(login)) > 0 {
-		fmt.Fprintf(w, "|login = %s\n", html.EscapeString(strings.ReplaceAll(login, "\n", "<br />")))
-	}
-	fmt.Fprintf(w, "|meet = %s\n|battle start = %s\n|battle end = %s\n|friendship max = %s\n|friendship event = %s\n",
-		html.EscapeString(strings.ReplaceAll(card.Meet(), "\n", "<br />")),
-		html.EscapeString(strings.ReplaceAll(card.BattleStart(), "\n", "<br />")),
-		html.EscapeString(strings.ReplaceAll(card.BattleEnd(), "\n", "<br />")),
-		html.EscapeString(strings.ReplaceAll(card.FriendshipMax(), "\n", "<br />")),
-		html.EscapeString(strings.ReplaceAll(card.FriendshipEvent(), "\n", "<br />")),
-	)
-	rebirth := card.RebirthEvent()
-	if len(strings.TrimSpace(rebirth)) > 0 {
-		fmt.Fprintf(w, "|rebirth = %s\n", html.EscapeString(strings.ReplaceAll(rebirth, "\n", "<br />")))
-	}
-
-	gevo, ok := evolutions["G"]
-	if !ok {
-		gevo, ok = evolutions["GA"]
-	}
-	if gevo != nil {
-		var awakenInfo *vc.CardAwaken
-		for idx, val := range vc.Data.Awakenings {
-			if gevo.ID == val.ResultCardID {
-				awakenInfo = &vc.Data.Awakenings[idx]
-				break
-			}
-		}
-		if awakenInfo != nil {
-			printAwakenMaterials(w, awakenInfo)
-		}
-	}
-
-	xevo, ok := evolutions["X"]
-	if !ok {
-		xevo, ok = evolutions["XA"]
-	}
-	if xevo != nil {
-		var rebirthInfo *vc.CardAwaken
-		for idx, val := range vc.Data.Rebirths {
-			if xevo.ID == val.ResultCardID {
-				rebirthInfo = &vc.Data.Rebirths[idx]
-				break
-			}
-		}
-		if rebirthInfo != nil {
-			printRebirthMaterials(w, rebirthInfo)
-		}
-	}
-
-	awl := make(vc.ArchwitchList, 0)
-	for _, evo := range evolutions {
-		aws := evo.ArchwitchesWithLikeabilityQuotes()
-		log.Printf("Evo AW records: %v", aws)
-		if len(aws) > 0 {
-			awl = append(awl, aws...)
-			log.Printf("Found %d AW records found for %s", len(awl), card.Name)
-		}
-	}
-	log.Printf("AW records: %v", awl)
-	if len(awl) > 0 {
-		aw := awl.Earliest()
-		log.Printf("AW %d found for card %s", aw.ID, card.Name)
-		for _, like := range aw.Likeability() {
-			fmt.Fprintf(w, "|likeability %d = %s\n",
-				like.Friendship,
-				html.EscapeString(strings.ReplaceAll(like.Likability, "\n", "<br />")),
-			)
-		}
-	}
-
-	if turnOverFrom != nil {
-		fmt.Fprintf(w, "|turnoverfrom = %s\n", turnOverFrom.Name)
-	} else if turnOverTo != nil {
-		fmt.Fprintf(w, "|turnoverto = %s\n", turnOverTo.Name)
-		fmt.Fprintf(w, "|availability = %s\n", avail)
-	} else {
-		fmt.Fprintf(w, "|availability = %s\n", avail)
-	}
-	io.WriteString(w, "}}")
+	io.WriteString(w, cardPage.String())
 
 	//Write out amalgamations here
 	if len(amalgamations) > 0 {
@@ -470,6 +268,7 @@ func CardDetailHandler(w http.ResponseWriter, r *http.Request) {
 	// show images here
 	io.WriteString(w, "<div style=\"float:left\">")
 	iconEvos := card.EvosWithDistinctImages(true)
+	lenEvoKeys := len(evokeys)
 	for _, k := range evokeys {
 		evo := evolutions[k]
 		// check if we should force non-H status
@@ -1000,356 +799,6 @@ func getNext(card *vc.Card) (next *vc.Card, nextName string) {
 		} else {
 			nextName = next.Name
 		}
-	}
-	return
-}
-
-func maxStats(evo *vc.Card, numOfEvos int) (atk, def, sol string) {
-	// stats := evo.EvoStandard()
-	// atk += fmt.Sprintf(" / %d", stats.Attack)
-	// def += fmt.Sprintf(" / %d", stats.Defense)
-	// sol += fmt.Sprintf(" / %d", stats.Soldiers)
-
-	// stats = evo.EvoMixed()
-	// atk += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", stats.Attack)
-	// def += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", stats.Defense)
-	// sol += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", stats.Soldiers)
-
-	// stats = evo.AmalgamationPerfect()
-	// atk += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Attack)
-	// def += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Defense)
-	// sol += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Soldiers)
-	// return
-
-	if evo.CardRarity() != nil && evo.CardRarity().MaxCardLevel == 1 && numOfEvos == 1 {
-		// only X cards have a max level of 1 and they don't evo
-		// only possible amalgamations like Philosopher's Stones
-		if evo.IsAmalgamation() {
-			stats := evo.AmalgamationPerfect()
-			atk = fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Attack)
-			def = fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Defense)
-			sol = fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", stats.Soldiers)
-		}
-		return
-	}
-	if evo.IsAmalgamation() {
-		stats := evo.EvoStandard()
-		atk = " / " + strconv.Itoa(stats.Attack)
-		def = " / " + strconv.Itoa(stats.Defense)
-		sol = " / " + strconv.Itoa(stats.Soldiers)
-		if strings.HasSuffix(evo.Rarity(), "LR") {
-			// print LR level1 static material amal
-			pStats := evo.AmalgamationPerfect()
-			if stats.NotEquals(pStats) {
-				if evo.PossibleMixedEvo() {
-					mStats := evo.EvoMixed()
-					if stats.NotEquals(mStats) {
-						atk += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Attack)
-						def += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Defense)
-						sol += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Soldiers)
-					}
-				}
-				lrStats := evo.AmalgamationLRStaticLvl1()
-				if lrStats.NotEquals(pStats) {
-					atk += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Attack)
-					def += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Defense)
-					sol += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Soldiers)
-				}
-				atk += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Attack)
-				def += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Defense)
-				sol += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Soldiers)
-			}
-		} else {
-			pStats := evo.AmalgamationPerfect()
-			if stats.NotEquals(pStats) {
-				atk += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Attack)
-				def += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Defense)
-				sol += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Soldiers)
-			}
-		}
-	} else if evo.EvolutionRank < 2 {
-		// not an amalgamation.
-		stats := evo.EvoStandard()
-		atk = " / " + strconv.Itoa(stats.Attack)
-		def = " / " + strconv.Itoa(stats.Defense)
-		sol = " / " + strconv.Itoa(stats.Soldiers)
-		pStats := evo.EvoPerfect()
-		if stats.NotEquals(pStats) {
-			var evoType string
-			if evo.PossibleMixedEvo() {
-				evoType = "Amalgamation"
-				mStats := evo.EvoMixed()
-				if stats.NotEquals(mStats) {
-					atk += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Attack)
-					def += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Defense)
-					sol += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Soldiers)
-				}
-			} else {
-				evoType = "Evolution"
-			}
-			atk += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Attack, evoType)
-			def += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Defense, evoType)
-			sol += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Soldiers, evoType)
-		}
-	} else {
-		// not an amalgamation, Evo Rank >=2 (Awoken cards or 4* evos).
-		stats := evo.EvoStandard()
-		atk = " / " + strconv.Itoa(stats.Attack)
-		def = " / " + strconv.Itoa(stats.Defense)
-		sol = " / " + strconv.Itoa(stats.Soldiers)
-		printedMixed := false
-		printedPerfect := false
-		if strings.HasSuffix(evo.Rarity(), "LR") {
-			// print LR level1 static material amal
-			if evo.PossibleMixedEvo() {
-				mStats := evo.EvoMixed()
-				if stats.NotEquals(mStats) {
-					printedMixed = true
-					atk += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Attack)
-					def += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Defense)
-					sol += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Soldiers)
-				}
-			}
-			pStats := evo.AmalgamationPerfect()
-			if stats.NotEquals(pStats) {
-				lrStats := evo.AmalgamationLRStaticLvl1()
-				if lrStats.NotEquals(pStats) {
-					atk += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Attack)
-					def += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Defense)
-					sol += fmt.Sprintf(" / {{tooltip|%d|LR 'Special' material Lvl-1, other materials Perfect Amalgamation}}", lrStats.Soldiers)
-				}
-				printedPerfect = true
-				atk += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Attack)
-				def += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Defense)
-				sol += fmt.Sprintf(" / {{tooltip|%d|Perfect Amalgamation}}", pStats.Soldiers)
-			}
-		}
-
-		if !strings.HasSuffix(evo.Rarity(), "LR") &&
-			evo.Rarity()[0] != 'G' && evo.Rarity()[0] != 'X' {
-			// TODO need more logic here to check if it's an Amalg vs evo only.
-			// may need different options depending on the type of card.
-			if evo.EvolutionRank == 4 {
-				//If 4* card, calculate 6 card evo stats
-				stats := evo.Evo6Card()
-				atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, 6)
-				def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, 6)
-				sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, 6)
-				if evo.Rarity() != "GLR" {
-					//If SR card, calculate 9 card evo stats
-					stats = evo.Evo9Card()
-					atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, 9)
-					def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, 9)
-					sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, 9)
-				}
-			}
-			//If 4* card, calculate 16 card evo stats
-			stats := evo.EvoPerfect()
-			var cards int
-			switch evo.EvolutionRank {
-			case 1:
-				cards = 2
-			case 2:
-				cards = 4
-			case 3:
-				cards = 8
-			case 4:
-				cards = 16
-			}
-			atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, cards)
-			def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, cards)
-			sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, cards)
-		}
-		if evo.EvoIsAwoken() || evo.EvoIsReborn() {
-			evo.EvoStandardLvl1() // just to print out the level 1 G stats
-			pStats := evo.EvoPerfect()
-			if stats.NotEquals(pStats) {
-				var evoType string
-				if !printedMixed && evo.PossibleMixedEvo() {
-					evoType = "Amalgamation"
-					mStats := evo.EvoMixed()
-					if stats.NotEquals(mStats) {
-						atk += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Attack)
-						def += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Defense)
-						sol += fmt.Sprintf(" / {{tooltip|%d|Mixed Evolution}}", mStats.Soldiers)
-					}
-				} else {
-					evoType = "Evolution"
-				}
-				if !printedPerfect {
-					atk += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Attack, evoType)
-					def += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Defense, evoType)
-					sol += fmt.Sprintf(" / {{tooltip|%d|Perfect %s}}", pStats.Soldiers, evoType)
-				}
-			}
-			awakensFrom := evo.AwakensFrom()
-			if awakensFrom == nil && evo.RebirthsFrom() != nil {
-				awakensFrom = evo.RebirthsFrom().AwakensFrom()
-			}
-			if awakensFrom != nil && awakensFrom.LastEvolutionRank == 4 {
-				//If 4* card, calculate 6 card evo stats
-				stats := evo.Evo6Card()
-				atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, 6)
-				def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, 6)
-				sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, 6)
-				if evo.MainRarity() != "LR" {
-					//If SR card, calculate 9 and 16 card evo stats
-					stats = evo.Evo9Card()
-					atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, 9)
-					def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, 9)
-					sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, 9)
-					stats = evo.EvoPerfect()
-					atk += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Attack, 16)
-					def += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Defense, 16)
-					sol += fmt.Sprintf(" / {{tooltip|%d|%d Card Evolution}}", stats.Soldiers, 16)
-				}
-			}
-		}
-	}
-	return
-}
-
-func printAwakenMaterials(w http.ResponseWriter, awakenInfo *vc.CardAwaken) {
-	if awakenInfo == nil {
-		return
-	}
-
-	fmt.Fprintf(w, "|awaken chance = %d\n", awakenInfo.Percent)
-
-	printAwakenMaterial(w, awakenInfo.Item(1), awakenInfo.Material1Count)
-	printAwakenMaterial(w, awakenInfo.Item(2), awakenInfo.Material2Count)
-	printAwakenMaterial(w, awakenInfo.Item(3), awakenInfo.Material3Count)
-	printAwakenMaterial(w, awakenInfo.Item(4), awakenInfo.Material4Count)
-	printAwakenMaterial(w, awakenInfo.Item(5), awakenInfo.Material5Count)
-
-}
-
-func printAwakenMaterial(w http.ResponseWriter, item *vc.Item, count int) {
-	if item == nil {
-		return
-	}
-	if strings.Contains(item.NameEng, "Crystal") {
-		fmt.Fprintf(w, "|awaken crystal = %d\n", count)
-	} else if strings.Contains(item.NameEng, "Orb") {
-		fmt.Fprintf(w, "|awaken orb = %d\n", count)
-	} else if strings.Contains(item.NameEng, "(L)") {
-		fmt.Fprintf(w, "|awaken l = %d\n", count)
-	} else if strings.Contains(item.NameEng, "(M)") {
-		fmt.Fprintf(w, "|awaken m = %d\n", count)
-	} else if strings.Contains(item.NameEng, "(S)") {
-		fmt.Fprintf(w, "|awaken s = %d\n", count)
-	} else {
-		fmt.Fprintf(w, "*******Unknown item: %s\n", item.NameEng)
-	}
-}
-
-func printRebirthMaterials(w http.ResponseWriter, awakenInfo *vc.CardAwaken) {
-	if awakenInfo == nil {
-		return
-	}
-
-	fmt.Fprintf(w, "|rebirth chance = %d\n", awakenInfo.Percent)
-
-	printRebirthMaterial(w, 1, awakenInfo.Item(1), awakenInfo.Material1Count)
-	printRebirthMaterial(w, 2, awakenInfo.Item(2), awakenInfo.Material2Count)
-	printRebirthMaterial(w, 3, awakenInfo.Item(3), awakenInfo.Material3Count)
-}
-
-func printRebirthMaterial(w http.ResponseWriter, matNum int, item *vc.Item, count int) {
-	if item == nil || count <= 0 {
-		return
-	}
-	fmt.Fprintf(w, "|rebirth item %d = %s\n", matNum, item.NameEng)
-	fmt.Fprintf(w, "|rebirth item %d count = %d\n", matNum, count)
-}
-
-func printWikiSkill(s *vc.Skill, ls *vc.Skill, evoMod string) (ret string) {
-	ret = ""
-	if s == nil {
-		return
-	}
-
-	if evoMod != "" {
-		evoMod += " "
-	}
-
-	lv10 := ""
-	skillLvl10 := ""
-	sName := s.Name
-	if ls == nil || ls.ID == s.ID {
-		// 1st skills only have lvl10, skill 2, 3, and thor do not.
-		if len(s.Levels()) == 10 {
-			skillLvl10 = html.EscapeString(strings.ReplaceAll(s.SkillMax(), "\n", "<br />"))
-			lv10 = fmt.Sprintf("\n|skill %slv10 = %s",
-				evoMod,
-				skillLvl10,
-			)
-		}
-	} else {
-		// handles Great DMG skills where the last evo is the max skill
-		sName = ls.Name
-		skillLvl10 = html.EscapeString(strings.ReplaceAll(ls.SkillMax(), "\n", "<br />"))
-		lv10 = fmt.Sprintf("\n|skill %slv10 = %s",
-			evoMod,
-			skillLvl10,
-		)
-	}
-
-	skillLvl1 := ""
-	if strings.HasSuffix(evoMod, "t ") {
-		skillLvl1 = s.FireMax() + " / 100% chance"
-	} else {
-		skillLvl1 = s.SkillMin()
-	}
-	skillLvl1 = html.EscapeString(strings.ReplaceAll(skillLvl1, "\n", "<br />"))
-
-	if skillLvl1 == skillLvl10 {
-		lv10 = ""
-	}
-
-	minturn := ""
-	costlvl1 := ""
-	costlvl10 := ""
-
-	if s.MinTurn > 0 {
-		minturn = fmt.Sprintf("\n|skill %smin turn = %d", evoMod, s.MinTurn)
-	}
-
-	if s.DefaultCost > 0 {
-		costlvl1 = fmt.Sprintf("\n|skill %slv1 cost = %d", evoMod, s.DefaultCost)
-	}
-
-	if s.MaxCost > 0 && s.MaxCost != s.DefaultCost {
-		costlvl10 = fmt.Sprintf("\n|skill %slv10 cost = %d", evoMod, s.MaxCost)
-	}
-
-	ret = fmt.Sprintf(`|skill %[1]s= %[2]s
-|skill %[1]slv1 = %[3]s%[4]s%[6]s%[7]s%[8]s
-|procs %[1]s= %[5]s
-`,
-		evoMod,
-		html.EscapeString(sName),
-		skillLvl1,
-		lv10,
-		s.ActivationString(),
-		minturn,
-		costlvl1,
-		costlvl10,
-	)
-
-	if s.EffectID == 36 {
-		// Random Skill
-		for k, v := range []int{s.EffectParam, s.EffectParam2, s.EffectParam3, s.EffectParam4, s.EffectParam5} {
-			rs := vc.SkillScan(v)
-			if rs != nil {
-				ret += fmt.Sprintf("|random %s%d = %s \n", evoMod, k+1, rs.FireMin())
-			}
-		}
-	}
-
-	// Check if the second skill expires
-	if s.Expires() {
-		ret += fmt.Sprintf("|skill %send = %v\n", evoMod, s.PublicEndDatetime)
 	}
 	return
 }
